+++ Picadito-de-Quidditch/src/main/Cazador.java
@@ -4,13 +4,13 @@ public class Cazador extends Jugador {
     private Integer punteria;
     private Integer fuerza;
 
-    Cazador(Integer punteria, Integer fuerza, Integer peso, Escoba escoba){
+    Cazador(Integer punteria, Integer fuerza, Double peso, Escoba escoba){
         super(peso, escoba);
         this.fuerza=fuerza;
         this.punteria=punteria;
     }
 
-    public Integer habilidad(){
-        return super.habilidad()+this.punteria*this.fuerza;
+    public Double habilidad(){
+        return super.habilidad()+(this.punteria*this.fuerza);
     }
 }
