+++ Picadito-de-Quidditch/src/main/Guardián.java
@@ -4,13 +4,13 @@ public class Guardián extends Jugador {
     private Integer nivelDeReflejos;
     private Integer fuerza;
 
-    Guardián(Integer nivelDeReflejos, Integer fuerza, Integer peso, Escoba escoba){
+    Guardián(Integer nivelDeReflejos, Integer fuerza, Double peso, Escoba escoba){
         super(peso, escoba);
         this.nivelDeReflejos=nivelDeReflejos;
         this.fuerza=fuerza;
     }
 
-    public Integer habilidad(){
+    public Double habilidad(){
         return super.habilidad()+this.nivelDeReflejos+this.fuerza;
     }
 }
